refact: 사이드바 메뉴 가로정렬 현상 수정
- innerWrapper에 메뉴 요소를 렌더링하는 방식이 가로 정렬의 원인임.
- innerWrapper 아래 contentWrapper를 넣고 메뉴 요소를 채워넣는 방식으로
  변경함.
- MVC로 작성하는 파일 이름을  SideBar로 변경함
- 기존 사이드바 파일을 SideBarLegacy로 변경함
